Refactor DnS and DINO baselines
