+++ vsc/baseline/localization.py
@@ -94,60 +94,3 @@ class VCSLLocalizationMaxSim(VCSLLocalization):
 class VCSLLocalizationCandidateScore(VCSLLocalization):
     def score(self, candidate: CandidatePair, match: Match, box, similarity) -> float:
         return candidate.score
-
-
-class VCSLLocalizationDnS(VCSLLocalizationMaxSim):
-    def __init__(
-        self,
-        model,
-        queries_fine,
-        refs_fine,
-        queries_coarse,
-        refs_coarse,
-        model_type,
-        device,
-        symmetric=True,
-        geometric_mean=True,
-        **kwargs,
-    ):
-        super().__init__(queries_coarse, refs_coarse, model_type, **kwargs)
-
-        self.queries_fine = queries_fine
-        self.refs_fine = refs_fine
-
-        self.sim_model = model
-        self.device = device
-
-        self.symmetric = symmetric
-        self.geometric_mean = geometric_mean
-
-    def _rescale_binaries(self, x):
-        if "bin" in self.sim_model.fg_type:
-            x = 2 * x - 1
-        return x
-
-    @torch.no_grad()
-    def similarity(self, candidate: CandidatePair):
-
-        query = self.queries_fine[candidate.query_id].feature
-        ref = self.refs_fine[candidate.ref_id].feature
-
-        query = torch.from_numpy(query).to(self.device).float()
-        ref = torch.from_numpy(ref).to(self.device).float()
-
-        query = self._rescale_binaries(query)
-        ref = self._rescale_binaries(ref)
-
-        sim = self.sim_model(query, ref)
-        if self.symmetric:
-            simT = self.sim_model(ref, query).mT
-            sim = (sim + simT) / 2.0
-        sim = sim.cpu().numpy()
-
-        if self.geometric_mean:
-            query = self.queries[candidate.query_id].feature
-            ref = self.refs[candidate.ref_id].feature
-
-            sim_cg = np.matmul(query, ref.T) + self.similarity_bias
-            sim = np.sqrt(sim.clip(1e-7) * sim_cg.clip(1e-7))
-        return sim
